Release story search includes done stories.
Added the includedone:true to the filter used to search for
release stories. This was obviously causing huge problems when
retrieving older releases. This is just hardcoded into the
url right now, since I can't think of any good reason why
it would make sense for this to be optional.